+++ src/main/java/se/technipelago/weather/vantagepro/DownloadController.java
@@ -24,7 +24,6 @@ import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 import org.apache.http.util.EntityUtils;
-import se.technipelago.pulsar.PulsarDataStore;
 import se.technipelago.weather.archive.ArchivePage;
 import se.technipelago.weather.archive.ArchiveRecord;
 import se.technipelago.weather.archive.CurrentRecord;
@@ -32,7 +31,7 @@ import se.technipelago.weather.archive.DataStore;
 import se.technipelago.opensensor.OpenSensorDataStore;
 import se.technipelago.weather.archive.RemoteDataStore;
 import se.technipelago.weather.archive.SqlDataStore;
-import se.technipelago.pulsar.PulsarDataStore;
+import nl.tudelft.davisstreaming.PulsarDataStore;
 
 import java.io.File;
 import java.io.FileInputStream;
